Make test/unit-mainnet for unit tests with mainnet
